Refactor for a more Pythonic implementation
